ext4_utils: Fix bad merge of EMMC wipe suppression
The order of the modules in the Makefile changed.

The Makefile items for eMMC wipe suppression should be
in the sections for target static and dynamic
libext4_utils.  These were the first two modules in the
file previously, but no longer are.

Move the eMMC wipe configuration handling to the proper
modules.

Change-Id: I86861a89824af04a3b08e75049580df28bd239f7

Conflicts:

	ext4_utils/Android.mk